fix: 835 - to evaluate is course stale on the client
