+++ src/core/base-layout/components/header/header.tsx
@@ -8,6 +8,7 @@ import { BurgerMenu } from './burger/burger';
 import { NavItem } from './nav-item/nav-item';
 import { ANCHORS, ROUTES } from '@/core/const';
 import { Course } from '@/entities/course';
+import { getActualData } from '@/shared/helpers/getActualData';
 import { Logo } from '@/shared/ui/logo';
 import { MobileView } from '@/widgets/mobile-view';
 import { SchoolMenu } from '@/widgets/school-menu';
@@ -28,6 +29,11 @@ export const Header = ({ courses }: HeaderProps) => {
   const pathname = usePathname();
   const headerAccentColor = pathname?.includes(ROUTES.MENTORSHIP) ? 'blue' : 'gray';
   const [color, setColor] = useState<HeaderAccentColor>(headerAccentColor);
+
+  const actualCourses = getActualData({
+    data: courses,
+    filterStale: false,
+  });
 
   const toggleMenu = () => {
     setMenuOpen((prev) => !prev);
@@ -67,7 +73,7 @@ export const Header = ({ courses }: HeaderProps) => {
         <Logo />
 
         <menu className={cx('mobile-menu', { open: isMenuOpen })} data-testid="mobile-menu">
-          <MobileView onClose={handleMenuClose} courses={courses} type="header" />
+          <MobileView onClose={handleMenuClose} courses={actualCourses} type="header" />
         </menu>
         <BurgerMenu isMenuOpen={isMenuOpen} toggleMenu={toggleMenu} />
 
@@ -87,7 +93,7 @@ export const Header = ({ courses }: HeaderProps) => {
           </NavItem>
           <NavItem label="Courses" href={ROUTES.COURSES}>
             <SchoolMenu>
-              {courses.map((course) => (
+              {actualCourses.map((course) => (
                 <SchoolMenu.Item
                   key={course.id}
                   icon={course.iconSmall}
