+++ src/widgets/courses/ui/courses.tsx
@@ -1,9 +1,8 @@
 import classNames from 'classnames/bind';
 
-import { type Course, CourseCard } from '@/entities/course';
 import { getCourses } from '@/entities/course/api/course-api';
-import { getActualData } from '@/shared/helpers/getActualData';
 import { WidgetTitle } from '@/shared/ui/widget-title';
+import CourseItems from '@/widgets/courses/ui/course-items';
 
 import styles from './courses.module.scss';
 
@@ -12,21 +11,12 @@ const cx = classNames.bind(styles);
 export const Courses = async () => {
   const courses = await getCourses();
 
-  const sortParams = {
-    data: courses,
-    filterStale: false,
-  };
-
-  const sortedCourses: Course[] = getActualData(sortParams);
-
   return (
     <section className={cx('container')} data-testid="all-courses">
       <div className={cx('content', 'courses-content')}>
         <WidgetTitle>All courses</WidgetTitle>
         <div className={cx('courses-list')}>
-          {sortedCourses.map((course) => {
-            return <CourseCard size="sm" key={course.id} {...course} />;
-          })}
+          <CourseItems courses={courses} />
         </div>
       </div>
     </section>
